requirements.txt
# Core
streamlit>=1.37
fastapi
uvicorn

# Data handling
pandas>=2.2
numpy
openpyxl>=3.1
XlsxWriter>=3.2

# Parsing & text utilities
beautifulsoup4>=4.12
lxml>=4.9
requests>=2.31
httpx
chardet
tqdm>=4.66
aiohttp
rapidfuzz>=3.9
tldextract>=5.1

# Search & external APIs
duckduckgo-search>=8.1.0
ddgs>=9.0.0

# Config & environment
python-dotenv>=1.0

sentence-transformers>=2.2.2
faiss-cpu>=1.7.4